No serving valid Atom XML
There are some quirks if you check it with the feed validator, like it
doesn’t understand what a figcaption tag is. The time and date
functions need to be cleaned up so that it isn’t set explicitly to
‘US/Pacific’ for any other person. Seemingly no good, reliable auto
detect for it, so it may need to be another settings.py arg.