Adds cProfiler flag, TypeKit support
Flag added to run python’s cProfile module with “-c” or “-cprofile”
performance for a blog with 122 posts to open is in the 2-3 second
range. Biggest hit is regex substitution. Typekit is now supported, if
an id isn’t present, it isn’t included in the head of the html files.
CSS styling adjustments.