Create yakbarber3.py
These are all the changes I had to make to get yakbarber to work in Python 3 for joe-steel.com. Any future development should be on this, but the original python script should be kept as reference until I can rule out any change there's an error. Of note, the new version of Beautiful Soup and the latest version of the python markdown library have significant changes both in arguments and in their output when compared to the old version.