Update yakbarber3.py
Not to jinx it but I think I finally figured out the footnotes issue? The thing was caching ALL INSTANCES of footnotes across ALL DOCUMENTS every time it ran so it would always pollute the output with random footnote garbage. This was apparently a design feature disabled with md.reset(). Ludicrous!